# specs/001-ai-consultation-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Car Service AI Assistant API
  version: 0.5.0
  description: >
    Полный операционный контракт, сгенерированный из runtime Express
    (`docs/api-route-inventory.json`). Схемы тел запросов — минимальные;
    источник истины по наличию path+method — инвентарь + `npm run openapi:check`.
servers:
  - url: http://localhost:3000/api
...
paths:
...
  "/consultations":
    get:
      tags: [Consultations]
      summary: Получить consultations
      operationId: get_consultations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Consultation' }
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Consultations]
      summary: Создать консультацию (клиент или гость)
      operationId: post_consultations
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsultationCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Consultation' }
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/abuse-challenge":
    get:
      tags: [Consultations]
      summary: Получить abuse-challenge
      operationId: get_consultations_abuse_challenge
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/context/active-templates":
    get:
      tags: [Consultations]
      summary: Получить active-templates
      operationId: get_consultations_context_active_templates
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/staff":
    get:
      tags: [Consultations]
      summary: Получить staff
      operationId: get_consultations_staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPage' }
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}":
    get:
      tags: [Consultations]
      summary: "Получить {sessionId}"
      operationId: get_consultations_sessionId
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/analyze-photo":
    post:
      tags: [Consultations]
      summary: Создать/выполнить analyze-photo
      operationId: post_consultations_sessionId_analyze_photo
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/claim":
    post:
      tags: [Consultations]
      summary: Создать/выполнить claim
      operationId: post_consultations_sessionId_claim
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/diagnosis-job":
    get:
      tags: [Consultations]
      summary: Получить diagnosis-job
      operationId: get_consultations_sessionId_diagnosis_job
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/export.pdf":
    get:
      tags: [Consultations]
      summary: Получить export.pdf
      operationId: get_consultations_sessionId_export_pdf
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/messages":
    post:
      tags: [Consultations]
      summary: Сообщение в консультацию
      operationId: post_consultations_sessionId_messages
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/messages/stream":
    post:
      tags: [Consultations]
      summary: Создать/выполнить stream
      operationId: post_consultations_sessionId_messages_stream
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/report":
    post:
      tags: [Consultations]
      summary: Создать/выполнить report
      operationId: post_consultations_sessionId_report
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/service-request":
    post:
      tags: [Consultations]
      summary: Создать/выполнить service-request
      operationId: post_consultations_sessionId_service_request
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/consultations/{sessionId}/service-request-guest":
    post:
      tags: [Consultations]
      summary: Создать/выполнить service-request-guest
      operationId: post_consultations_sessionId_service_request_guest
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error: { type: string }
        code: { type: string }
    ConsultationCreate:
      type: object
      additionalProperties: true
      properties:
        serviceCategoryId: { type: string, format: uuid }
    Consultation:
      type: object
      properties:
        id: { type: string, format: uuid }
        guestToken: { type: string, nullable: true }
        status: { type: string }
...
    CursorPage:
      type: object
      properties:
        items: { type: array, items: { type: object } }
        nextCursor: { type: string, nullable: true }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }